fix: bug in bulk_save ep for cdt
There was an issue when saveable was deleted all existing condition groups

+++ src/django_app/tables/services/graph_bulk_save_service/saveables.py
@@ -237,7 +237,7 @@ class ClassificationDecisionTableNodeSaveable:
             else s.update(s.instance, validated)
         )
 
-        if self._instance is not None:
+        if self._instance is not None and self._condition_groups_data is not None:
             ClassificationConditionGroup.objects.filter(
                 classification_decision_table_node=node
             ).delete()
